Refactor stateless components to use function syntax
