+++ src/App/screens/Instructor/components/ListOfLessons/index.js
@@ -1,0 +1,37 @@
+import React from 'react'
+import ReactPaginate from 'react-paginate'
+import LessonListItem from './components/LessonListItem'
+
+const ListOfLessons = (
+  lessons=[],
+  selectedLesson,
+  selectLesson,
+  total,
+  handlePageClick,
+  currentPage,
+) => (
+  <main className="mw6">
+    <div id="react-paginate">
+      <ReactPaginate 
+        previousLabel={"previous"}
+        nextLabel={"next"}
+        breakLabel={<a href="">...</a>}
+        pageNum={Math.ceil(total / 5)}
+        marginPagesDisplayed={1}
+        pageRangeDisplayed={3}
+        initialSelected={currentPage}
+        clickCallback={(page) => {
+          handlePageClick(page.selected+1)
+        }}
+      />
+    </div>
+    {lessons.map((lesson) => (
+      <LessonListItem selectedLesson={selectedLesson}
+        lesson={lesson}
+        selectLesson={selectLesson}
+      />
+    ))}
+  </main>
+)
+
+export default ListOfLessons
