--- src/App/screens/Instructor/screens/LessonFeedback/index.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-class LessonFeedback extends React.Component {
-  render() {
-    return (
-      <h1>Lesson Feedback</h1>
-    )
-  }
-}
-
-export default LessonFeedback
